utils: split sync retry wrapper
