+++ kubernetes/aio/utils/__init__.py
@@ -22,6 +22,5 @@ from .retry import (Backoff, DEFAULT_BACKOFF, DEFAULT_RETRY,
                     DEFAULT_RETRY_AFTER_BACKOFF, async_on_error,
                     async_on_retry_after_error, async_retry_on_conflict,
                     is_conflict, is_retry_after_response,
-                    is_too_many_requests, on_error, on_retry_after_error,
-                    retry_after_backoff, retry_after_max_retries,
-                    retry_on_conflict, retry_after_seconds)
+                    is_too_many_requests, retry_after_backoff,
+                    retry_after_max_retries, retry_after_seconds)
